Check for repetition of empty-matching expression
Add check for repetition of expressions that may match an empty string,
and repetition of zero-width assertions.
The former is experimental but seems useful since such expressions can be
very slow on the backtracking regexp engine in Emacs.
The latter is just to catch silly mistakes, which it occasionally does.