Check for sequence of subsuming repetitions
This check finds things like [ab]+b?, where one term (b? in this case)
can be removed because it is subsumed by the other. This can indicate
an error in the regexp, but also potentially slow matches because
there are multiple ways to match strings (consider a*a*).

The exact rules are: both terms are repetitions, one operand matches a
superset of the other, the subsumed repetition can match 0 times, and
the subsuming repetition can match infinitely many times.