Recognise implicit gap in ASCII-raw ranges
Character ranges between ASCII and raw bytes, such as "[a-\xc0]", do
not include the intervening codes 80..3fff7f. That is, "[a-\xc0]"
really means "[a-\x7f\x80-\xc0]" and does not match any non-ASCII
Unicode character. This is also true for skip-sets.

This change does two things: xr-lint no longer complains about
"[a-\xc0λ]" since the λ is in the gap not matched by the range, and
the conversion to rx now represents the gap explicitly:
(any "a-\x7f\x80-\xc0" "λ").