Translate \w and \W to (syntax word) and (not (syntax word))
We used to translate `\w` to `wordchar` which rx renders as
`[[:word:]]`, but they are not quite equivalent: `[[:word:]]` doesn't
heed syntax properties but `\w` does. Translating `\W` to
`not-wordchar` is correct but it is confusingly named so that is
changed as well.

Rx translates `(syntax word)` to `\sw` but maybe it could use the
one byte shorter `\w`, to complete the roundtrip.